feat: extract pipeline functions from scripts/ into leadforge.pipelines.build_v5
Moves reusable data transformation functions (subsample, inject_missingness,
derive_binary_features, cap_expected_acv, rename_and_select, boost_leakage_trap)
and their constants into a proper package module. The script becomes a thin CLI
wrapper, and tests import directly from the package instead of using importlib.

Closes #29
