+++ leadforge/pipelines/__init__.py
@@ -1,0 +1,1 @@
+"""Pipeline modules for building derived datasets from leadforge bundles."""
