+++ tests/scripts/test_build_v5_snapshot.py
@@ -1,36 +1,27 @@
-"""Tests for scripts/build_v5_snapshot.py pipeline functions."""
+"""Tests for leadforge.pipelines.build_v5 pipeline functions."""
 
 from __future__ import annotations
-
-import importlib.util
-from pathlib import Path
 
 import numpy as np
 import pandas as pd
 import pytest
 
-# ---------------------------------------------------------------------------
-# Import the script module (not in a package, so use importlib)
-# ---------------------------------------------------------------------------
-_SCRIPT_PATH = Path(__file__).resolve().parents[2] / "scripts" / "build_v5_snapshot.py"
+from leadforge.pipelines.build_v5 import (
+    ACV_CAP,
+    ACV_FLOOR,
+    FINAL_COLUMNS,
+    RENAME_MAP,
+    boost_leakage_trap,
+    cap_expected_acv,
+    derive_binary_features,
+    inject_missingness,
+    rename_and_select,
+    subsample,
+)
 
-spec = importlib.util.spec_from_file_location("build_v5_snapshot", _SCRIPT_PATH)
-assert spec is not None
-assert spec.loader is not None
-build_v5 = importlib.util.module_from_spec(spec)
-spec.loader.exec_module(build_v5)
-
-# Re-export for convenience
-subsample = build_v5.subsample
-inject_missingness = build_v5.inject_missingness
-derive_binary_features = build_v5.derive_binary_features
-cap_expected_acv = build_v5.cap_expected_acv
-rename_and_select = build_v5.rename_and_select
-boost_leakage_trap = build_v5.boost_leakage_trap
-ACV_FLOOR = build_v5.ACV_FLOOR
-ACV_CAP = build_v5.ACV_CAP
-_FINAL_COLUMNS = build_v5._FINAL_COLUMNS
-_RENAME_MAP = build_v5._RENAME_MAP
+# Backwards-compatible aliases used in tests below.
+_FINAL_COLUMNS = FINAL_COLUMNS
+_RENAME_MAP = RENAME_MAP
 
 
 # ---------------------------------------------------------------------------
